backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
pyjwt==2.8.0
bcrypt==4.1.3
pydantic==2.7.0
groq>=0.11.0
python-dotenv==1.0.1
cryptography==42.0.8
razorpay==1.4.1
stripe>=9.0.0
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0
httpx>=0.27.0
firebase-admin>=6.5.0
slowapi>=0.1.9
email-validator>=2.1.0
